requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
requests==2.32.0
urllib3==2.2.0
streamlit==1.38.0
sentence-transformers==3.1.0
faiss-cpu==1.8.0
numpy==1.26.4
torch>=2.0.0
python-dotenv==1.0.1
